--- client/src/features/Products/components/ProductModal.jsx
@@ -1,9 +1,0 @@
-import PropTypes from 'prop-types';
-
-function ProductModal(props) {
-  return <div>ProductModal</div>;
-}
-
-ProductModal.propTypes = {};
-
-export default ProductModal;
