+++ client/src/features/Products/components/index.jsx
@@ -3,7 +3,6 @@ export { default as ProductDisplay } from './ProductDisplay';
 export { default as ProductFilters } from './ProductFilters';
 export { default as ProductForm } from './ProductForm';
 export { default as ProductGrid } from './ProductGrid';
-export { default as ProductModal } from './ProductModal';
 export { default as ProductPage } from './ProductPage';
 export { default as ProductTable } from './ProductTable';
